requirements.txt
dnspython==2.8.0
numpy==2.4.4
pandas==3.0.2
pymongo==4.16.0
python-dateutil==2.9.0.post0
six==1.17.0
python-dotenv==1.0.0
